backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
motor==3.5.1
feedparser==6.0.11
sumy==0.11.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT[crypto]==2.9.0
APScheduler==3.10.4
email-validator==2.2.0

